Move mint header to mint component - Also fetch epoch and use countdown rather than a real value
